Rename func_utils to micro
